+++ app/utils/constants.ts
@@ -1,0 +1,38 @@
+export const LANGUAGE_MODEL_OPTIONS = [
+  {
+    label: "Dall-E",
+    value: "dall-e",
+  },
+  {
+    label: "Stable Diffusion XL",
+    value: "stable-diffusion-xl",
+  },
+  {
+    label: "Stable Diffusion 1.5",
+    value: "stable-diffusion-1-5",
+  },
+  {
+    label: "Stable Diffusion 2.1",
+    value: "stable-diffusion-2-1",
+  },
+];
+
+export const STABLE_DIFFUSION_IMAGE_PRESETS = [
+  { value: "3d-model", label: "3d-model" },
+  { value: "analog-film", label: "analog-film" },
+  { value: "anime", label: "anime" },
+  { value: "cinematic", label: "cinematic" },
+  { value: "comic-book", label: "comic-book" },
+  { value: "digital-art", label: "digital-art" },
+  { value: "enhance", label: "enhance" },
+  { value: "fantasy-art", label: "fantasy-art" },
+  { value: "isometric", label: "isometric" },
+  { value: "line-art", label: "line-art" },
+  { value: "low-poly", label: "low-poly" },
+  { value: "modeling-compound", label: "modeling-compound" },
+  { value: "neon-punk", label: "neon-punk" },
+  { value: "origami", label: "origami" },
+  { value: "photographic", label: "photographic" },
+  { value: "pixel-art", label: "pixel-art" },
+  { value: "tile-texture", label: "tile-texture" },
+];
